requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
opencv-python>=4.7.0
Pillow>=9.0.0
